Add counting and returning count
At this point the system will send the file's contents to the child
process, the child process will count the words, and send the size back
to the parent. Additionally, The print statements are being made.

Note: There seems to be a issue somewhere with the amount of words being
counted or sent. The next stage is to refactor this program into smaller
more easily debugged functions - and write tests for them.