Separate Pipelining process from main
The pipeline creation and error checking have been moved into it's own
.c and .h files. This is an attempt to simplify functions for testing
purposes.

I discovered that libraries not added as submodules would not load into
the path. To pass this we must add them as submodules and tell cmake
where to build them:

```
add_submodule(Directory /path/to/build/directory)
```